refactor(engineer_loop/selection): extract tests to siblings (#1266)
Partial: selection.rs 1352 -> 759 LOC. Tests split into two new
siblings tests_selection_inline_a.rs and tests_selection_inline_b.rs
(296+357 LOC) since existing tests_selection.rs and tests_selection_extra.rs
were already near the limit. Source split deferred.
